refactor(default-layout): yak shaving - separate hasSearchableTerms and selectors, move into utils
