+++ packages/@sanity/default-layout/src/navbar/search/components/SearchPopover.tsx
@@ -1,14 +1,14 @@
+import {useId} from '@reach/auto-id'
 import {Box, Card, Flex, Portal, Theme, useClickOutside, useLayer} from '@sanity/ui'
 import React, {useCallback, useEffect, useRef, useState} from 'react'
 import FocusLock from 'react-focus-lock'
 import styled, {css} from 'styled-components'
-import {useId} from '@reach/auto-id'
 import {POPOVER_INPUT_PADDING, POPOVER_MAX_HEIGHT, POPOVER_MAX_WIDTH} from '../constants'
 import {CommandListProvider} from '../contexts/commandList'
 import {useSearchState} from '../contexts/search'
-import {hasSearchableTerms} from '../contexts/search/selectors'
 import {useMeasureSearchResultsIndex} from '../hooks/useMeasureSearchResultsIndex'
 import {useSearchHotkeys} from '../hooks/useSearchHotkeys'
+import {hasSearchableTerms} from '../utils/hasSearchableTerms'
 import {RecentSearches} from './RecentSearches'
 import {SearchHeader} from './SearchHeader'
 import {SearchResults} from './SearchResults'
