+++ packages/@sanity/default-layout/src/navbar/search/utils/hasSearchableTerms.ts
@@ -1,0 +1,5 @@
+import type {SearchTerms} from '@sanity/base'
+
+export function hasSearchableTerms(terms: SearchTerms): boolean {
+  return terms.query.trim() !== '' || !!terms.types.length
+}
